requirements.txt
beautifulsoup4==4.14.2
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
packaging==25.0
pdfminer.six==20251107
pip3-autoremove==2.0.1
pycparser==2.23
soupsieve==2.8
typing_extensions==4.15.0
